Move projections into their own module
xrdutil was getting too large, and we were running into circular import
issues by referencing too many other parts of the code.

Moving projections into their own module helps prevent circular import
issues since they will not be imported unless they are actually needed.
